+++ shared/src/state-migrations/10-rename-delete-transfer-action.ts
@@ -3,7 +3,7 @@ import type { Migration } from './migration-functions';
 export const renameDeleteTransferAction10: Migration = {
     action: (_intermediaryState, action) => {
         const typedAction = action as { type: string };
-        if (typedAction?.type === '[Transfer] delete transfer') {
+        if (typedAction.type === '[Transfer] delete transfer') {
             typedAction.type = '[Transfer] Finish transfer';
         }
         return true;