+++ shared/src/state-migrations/30-report-treatment-status-changes.ts
@@ -28,6 +28,8 @@ export const reportTreatmentStatusChanges30: Migration = {
             const typedAction = action as {
                 behaviorState: BehaviorStateStub;
             };
+            // Disabled due to incomplete typings in the migration
+            // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
             if (typedAction.behaviorState.type === 'reportBehavior') {
                 migrateReportBehavior(typedAction.behaviorState);
             }
@@ -57,6 +59,8 @@ export const reportTreatmentStatusChanges30: Migration = {
         );
 
         Object.values(typedState.radiograms).forEach((radiogram) => {
+            // Disabled due to incomplete typings in the migration
+            // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
             if (radiogram.type === 'treatmentStatusRadiogram') {
                 radiogram.treatmentStatusChanged = false;
             }
@@ -67,6 +71,8 @@ export const reportTreatmentStatusChanges30: Migration = {
 function migrateSimulatedRegion(simulatedRegion: SimulatedRegionStub) {
     simulatedRegion.behaviors.forEach((behavior) => {
         {
+            // Disabled due to incomplete typings in the migration
+            // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
             if (behavior.type === 'reportBehavior') {
                 migrateReportBehavior(behavior);
             }