+++ shared/src/utils/uuid-set.ts
@@ -1,5 +1,5 @@
 import type { UUID } from '.';
 
-export interface UUIDSet {
-    readonly [key: UUID]: true;
-}
+export type UUIDSet = {
+    readonly [key in UUID]: true;
+};