+++ CHANGELOG.md
@@ -20,6 +20,14 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
     -   Treatment status changes.
     -   Completion of the transfer of patients of one category
         -   Either in one region or in all regions managed by one behavior
+    -   Vehicle transfer, loading, unloading and occupation changes.
+    -   Visible status changes of patients.
+    -   Treating personnel changes for patients.
+-   By clicking on a log entry or a chart, a marker will be shown in the chart at that time. The log entry list scrolls to that time.
+
+### Fixed
+
+-   Errors in reduction of the tick actions no longer crash the backend.
 
 ## [0.6.0] - 2023-05-17